Scenario implementation and different shopkeepers
Using money to buy items.
New Items

+++ Assets/Scripts/ItemScriptable.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class ItemScriptable : ScriptableObject
 {
     public string itemName;
-    public float itemPrice;
+    public int itemPrice;
     public Sprite itemIcon;
     public ItemType itemType;
